+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/common/configuration/integration/ExtraHNNConfiguration.kt
@@ -1,0 +1,16 @@
+package site.siredvin.peripheralworks.common.configuration.integration
+
+import site.siredvin.peripheralworks.common.configuration.IntegrationConfiguration
+import net.neoforged.neoforge.common.ModConfigSpec as ForgeConfigSpec
+
+object ExtraHNNConfiguration : IntegrationConfiguration {
+    override val modID: String = "extrahnn"
+    override val name: String = "extrahnn"
+    private var enabledConfig: ForgeConfigSpec.BooleanValue? = null
+    val enabled: Boolean
+        get() = enabledConfig.getOrDefault(true)
+
+    override fun addToConfig(builder: ForgeConfigSpec.Builder) {
+        enabledConfig = builder.comment("Enables loot selection and model item details").define("enabled", true)
+    }
+}